+++ CHANGELOG.md
@@ -7,13 +7,19 @@
 - Improve compiler error message for missing Swift declarations due to APPLICATION_EXTENSION_API_ONLY (#4603)
 - Add file IO tracking wrapper for Swift.Data (#4605)
 
+### Features
+
+- Show session replay options as replay tags (#4639)
+
 ### Fixes
 
 - `SentrySdkInfo.packages` should be an array (#4626)
+- Use the same SdkInfo for envelope header and event (#4629)
 
 ### Internal
 
 - Remove loading `integrations` names from `event.extra` (#4627)
+- Add Hybrid SDKs API to add extra SDK packages (#4637)
 
 ## 8.42.0-beta.2